+++ src/main/java/com/codecool/seriesapp/controller/StaffController.java
@@ -1,6 +1,5 @@
 package com.codecool.seriesapp.controller;
 
-import com.codecool.seriesapp.model.generated.Series;
 import com.codecool.seriesapp.model.generated.people.People;
 import com.codecool.seriesapp.service.PeopleSearchApi;
 import com.codecool.seriesapp.service.SeriesApiService;
